Add files via upload
Each team is required to compose 5-10 seconds of music of your own choice based on the
instructions given in Appendix A and accomplish the following tasks:
1. Plot the time and frequency domain components of the music you created;
2. Explain the output of your graphs according to your design.
3. Plot the spectrogram of the music you created and explain your findings from the
generated spectrogram.
Extra credit challenges (15 points):
a. Add white Gaussian noise to your music and analyze and plot the output music in both time
and frequency domains. Explain the signal to noise ratio of your output music.
b. Design a filter to remove noise
c. Analyze and plot the music in both time and frequency domain.